Add explicit type conversions where necessary: base/
These are cases that are implicitly narrowing today, and must do so
explicitly in order to enable -Wc++11-narrowing.  No behavior change
intended.

Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2947112
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#890831}

+++ base/profiler/metadata_recorder_unittest.cc
@@ -198,7 +198,8 @@ TEST(MetadataRecorderTest, ReclaimInactiveSlots) {
   // Fill up the metadata map.
   for (size_t i = 0; i < MetadataRecorder::MAX_METADATA_COUNT; ++i) {
     recorder.Set(i, absl::nullopt, i);
-    items_set.insert(MetadataRecorder::Item{i, absl::nullopt, i});
+    items_set.insert(
+        MetadataRecorder::Item{i, absl::nullopt, static_cast<int64_t>(i)});
   }
 
   // Remove every fourth entry to fragment the data.
@@ -206,13 +207,15 @@ TEST(MetadataRecorderTest, ReclaimInactiveSlots) {
   for (size_t i = 3; i < MetadataRecorder::MAX_METADATA_COUNT; i += 4) {
     recorder.Remove(i, absl::nullopt);
     ++entries_removed;
-    items_set.erase(MetadataRecorder::Item{i, absl::nullopt, i});
+    items_set.erase(
+        MetadataRecorder::Item{i, absl::nullopt, static_cast<int64_t>(i)});
   }
 
   // Ensure that the inactive slots are reclaimed to make room for more entries.
   for (size_t i = 1; i <= entries_removed; ++i) {
     recorder.Set(i * 100, absl::nullopt, i * 100);
-    items_set.insert(MetadataRecorder::Item{i * 100, absl::nullopt, i * 100});
+    items_set.insert(MetadataRecorder::Item{i * 100, absl::nullopt,
+                                            static_cast<int64_t>(i * 100)});
   }
 
   MetadataRecorder::ItemArray items_arr;
